requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0

beautifulsoup4==4.13.4
blis==1.3.0
catalogue==2.0.10
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
cssselect==1.3.0
cymem==2.0.11
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
eval_type_backport==0.2.2
exceptiongroup==1.3.0
faiss-cpu==1.11.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
h11==0.16.0
hf-xet==1.1.6rc2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
iniconfig==2.1.0
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-core==0.3.72
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.4.8
language_data==1.3.0
lxml==6.0.0
lxml_html_clean==0.4.2
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
murmurhash==1.0.13
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.6
ollama==0.5.1
orjson==3.11.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
preshed==3.0.10
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
sentence-transformers==5.0.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.3.0
together==1.5.21
tokenizers==0.21.2
tomli==2.2.1
torch==2.7.1
tqdm==4.67.1
transformers==4.53.3
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
yarl==1.20.1
zstandard==0.23.0
